:bug: s16: Fix for client loading of youtube content
Can't access the youtube api key securely otherwise
